# perf(swc_common):split-hygine-data