<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**
1. Split `HygineData` into two parts. `HygineData.maks` and (`HygineData.syntax_contexts`, `HygineData.markings`), are always used exclusively, using two `Mutex` could reduce some unnecessary locking. e.g. It is not necessary to lock  (`HygineData.syntax_contexts`, `HygineData.markings`) when invoke `Mark::refresh() `, vise versa.
This give us about 10% performance improvement in `transform_base/parallel`
**before**: 
```
es/base/parallel/resolver/typescript                                                                          
                        time:   [13.508 s 13.631 s 13.743 s]
Found 1 outliers among 10 measurements (10.00%)
```

**after**:
```
es/base/parallel/resolver/typescript                                                                          
                        time:   [12.107 s 12.247 s 12.382 s]
```
<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

**BREAKING CHANGE:**

<!--
If this PR introduces a breaking change, it must contain a notice for it to be included in the CHANGELOG. Add description or remove entirely if not breaking.

You may need to update `MIGRATION.md` for the breaking changes.
-->

**Related issue (if exists):**